Move getUser, putUser and updateUser
