+++ app/src/main/java/ch/epfl/sdp/appart/database/firestoreservicehelpers/FirestoreCardHelper.java
@@ -1,0 +1,4 @@
+package ch.epfl.sdp.appart.database.firestoreservicehelpers;
+
+public class FirestoreCardHelper {
+}
